requirements.txt
affine==2.4.0
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
django-admin-adminlte==1.0.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
fiona==1.10.1
fonttools==4.55.3
geopandas==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
rasterio==1.4.3
rasterstats==0.20.0
shapely==2.0.6
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
gunicorn==20.1.0
whitenoise==6.2.0